% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epispot.R
\name{epispot}
\alias{epispot}
\title{Fit sparse multivariate regression models using variational inference.}
\usage{
epispot(
  Y,
  X,
  p0,
  V,
  list_blocks = NULL,
  list_hyper = NULL,
  list_init = NULL,
  user_seed = NULL,
  tol = 0.001,
  adaptive_tol_em = FALSE,
  maxit = 1000,
  anneal = NULL,
  anneal_vb_em = NULL,
  save_hyper = FALSE,
  save_init = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{Y}{Response data matrix of dimension n x d, where n is the number of
samples and d is the number of response variables.}

\item{X}{Input matrix of dimension n x p, where p is the number of candidate
predictors. \code{X} cannot contain NAs. No intercept must be supplied.}

\item{p0}{Vector of size 2 whose entries are the prior expectation and 
variance of the number of predictors associated with each response.
Must be \code{NULL} if \code{list_init} and \code{list_hyper} are both 
non-\code{NULL}.}

\item{V}{Annotation matrix of dimension p x r, where r is the number of
variables representing external information on the candidate predictors
which may make their selection more or less likely. \code{NULL} if no such
information.}

\item{list_blocks}{An object of class "\code{blocks}" containing settings for
parallel inference on a partitioned predictor space. Must be filled using
the \code{\link{set_blocks}} function or must be \code{NULL} for no
partitioning.}

\item{list_hyper}{An object of class "\code{hyper}" containing the model
hyperparameters. Must be filled using the \code{\link{set_hyper}}
function or must be \code{NULL} for default hyperparameters.}

\item{list_init}{An object of class "\code{init}" containing the initial
variational parameters. Must be filled using the \code{\link{set_init}}
function or be \code{NULL} for a default initialization.}

\item{user_seed}{Seed set for reproducible default choices of hyperparameters
(if \code{list_hyper} is \code{NULL}) and initial variational parameters
(if \code{list_init} is \code{NULL}). Default is \code{NULL}, no
seed set.}

\item{tol}{Tolerance for the stopping criterion.}

\item{adaptive_tol_em}{Boolean indicating whether the tolerance for the 
within-EM variational runs should be controlled adaptively depending on the
EM-convergence status.}

\item{maxit}{Maximum number of iterations allowed.}

\item{anneal}{Parameters for annealing scheme. Must be a vector whose first
entry is sets the type of ladder: 1 = geometric spacing, 2 = harmonic
spacing or 3 = linear spacing, the second entry is the initial temperature,
and the third entry is the ladder size. If \code{NULL} (default), no
annealing is performed.}

\item{anneal_vb_em}{Parameters for annealing scheme for the internal runs of the 
variational EM algorithm.}

\item{save_hyper}{If \code{TRUE}, the hyperparameters used for the model are
saved as output.}

\item{save_init}{If \code{TRUE}, the initial variational parameters used for
the inference are saved as output.}

\item{verbose}{If \code{TRUE}, messages are displayed during execution.}
}
\value{
An object of class "\code{vb}" containing the following variational
  estimates and settings:
 \item{gam_vb}{Posterior inclusion probability matrix of dimension p x d.
               Entry (s, t) corresponds to the posterior probability of
               association between candidate predictor s and response t.}
 \item{mu_c_vb}{Vector of size r, where entry l contains the overall effect 
                of annotation l on the
                probabilities of associations.\code{NULL} if \code{V} is
                \code{NULL}.}
 \item{mu_rho_vb}{Vector of length d containing the posterior mean of rho.
                  Entry t controls the proportion of predictors associated
                  with response t.}
 \item{mu_theta_vb}{Vector of length p containing the posterior mean of
                    theta. Entry s corresponds to the propensity of candidate
                    predictor s to be included in the model.}
 \item{om}{Vector of length p containing the posterior mean of omega.
              Entry s controls the proportion of responses associated with
              candidate predictor s. NULL if \code{V} is non-\code{NULL}.}
 \item{zeta_vb}{Posterior inclusion probability vector of size r for the
                annotation variables. \code{NULL} if \code{V} is \code{NULL}.}
 \item{converged}{A boolean indicating whether the algorithm has converged
                  before reaching \code{maxit} iterations.}
 \item{it}{Final number of iterations.}
 \item{lb_opt}{Optimized variational lower bound for the marginal
               log-likelihood (ELBO).}
 \item{diff_lb}{Difference in ELBO between the last and penultimate
                iterations. This may be a useful diagnostic information when
                convergence has not been reached before \code{maxit}.}
 \item{p0}{Vector of length 2 defining the applied sparsity control.}
 \item{rmvd_cst_x}{Vectors containing the indices of constant
                               variables in \code{X} removed
                               prior to the analysis.}
 \item{rmvd_coll_x}{Vectors containing the indices of variables
                                 in \code{X} removed prior
                                 to the analysis because collinear to other
                                 variables. The entry name indicates the
                                 corresponding variable kept in the analysis
                                 (i.e., that causing the collinearity for the
                                 entry in question).}
 \item{list_hyper, list_init}{If \code{save_hyper}, resp. \code{save_init},
                              \code{TRUE}, hyperparameters, resp. initial
                              variational parameters, used for inference are
                              saved as output.}
 \item{...}{Other specific outputs are possible depending on the model used.}
}
\description{
Variational approximation procedure fitting sparse multivariate regression
models for combined selection of predictors and associated responses in
high-dimensional set-ups. Dependence across responses linked to the same
predictors is captured through the model hierarchical structure.
}
\details{
The optimization is made using efficient block coordinate ascent schemes, for
which convergence is ensured as the objective (elbo) is multiconcave
for the selected blocks, i.e., it is concave in each block of parameters
whose updates are made simultaneously, see Wu et al. (reference Section
below).

The continuous response variables in \code{Y} (if any) will be centered
before application of the variational algorithm, and the candidate predictors
in \code{X} will be standardized.
}
\examples{
seed <- 123; set.seed(seed)

###################
## Simulate data ##
###################

## Examples using small problem sizes:
##
n <- 200; p <- 250; p0 <- 25; d <- 30; d0 <- 25; q <- 3; r <- 3

## Candidate predictors (subject to selection)
##
# Here we simulate common genetic variants (but any type of candidate
# predictors can be supplied).
# 0 = homozygous, major allele, 1 = heterozygous, 2 = homozygous, minor allele
#
X_act <- matrix(rbinom(n * p0, size = 2, p = 0.25), nrow = n)
X_inact <- matrix(rbinom(n * (p - p0), size = 2, p = 0.25), nrow = n)

shuff_x_ind <- sample(p)
X <- cbind(X_act, X_inact)[, shuff_x_ind]

bool_x_act <- shuff_x_ind <= p0

pat_act <- beta <- matrix(0, nrow = p0, ncol = d0)
pat_act[sample(p0*d0, floor(p0*d0/5))] <- 1
beta[as.logical(pat_act)] <-  rnorm(sum(pat_act))

## Covariates (not subject to selection)
##
Z <- matrix(rnorm(n * q), nrow = n)

alpha <-  matrix(rnorm(q * d), nrow = q)

## Gaussian responses
##
Y_act <- matrix(rnorm(n * d0, mean = X_act \%*\% beta, sd = 0.5), nrow = n)
Y_inact <- matrix(rnorm(n * (d - d0), sd = 0.5), nrow = n)
shuff_y_ind <- sample(d)
Y <- cbind(Y_act, Y_inact)[, shuff_y_ind] + Z \%*\% alpha

## Binary responses
##
Y_bin <- ifelse(Y > 0, 1, 0)

## Informative annotation variables
##
V <- matrix(rnorm(p * r), nrow = p)
V[bool_x_act, ] <- rnorm(p0 * r, mean = 2)


########################
## Infer associations ##
########################

## Continuous responses
##

# No covariate
#
vb_g <- epispot(Y = Y, X = X, p0 = p0, link = "identity", user_seed = seed)

# With covariates
#
vb_g_z <- epispot(Y = Y, X = X, p0 = p0,  Z = Z, link = "identity",
                user_seed = seed)

# With external annotation variables
#
vb_g_v <- epispot(Y = Y, X = X, p0 = p0, Z = Z, V = V, s2 = 0.1, 
                  link = "identity", user_seed = seed)

## Binary responses
##
vb_logit <- epispot(Y = Y_bin, X = X, p0 = p0, Z = Z, link = "logit",
                  user_seed = seed)

vb_probit <- epispot(Y = Y_bin, X = X, p0 = p0, Z = Z, link = "probit",
                   user_seed = seed)

## Mix of continuous and binary responses
##
Y_mix <- cbind(Y, Y_bin)
ind_bin <- (d+1):(2*d)

vb_mix <- epispot(Y = Y_mix, X = X, p0 = p0, Z = Z, link = "mix",
                ind_bin = ind_bin, user_seed = seed)

}
\references{
H. Ruffieux, A. C. Davison, J. Hager, I. Irincheeva. Efficient inference for
  genetic association studies with multiple outcomes. Biostatistics, 2017.

Y. Xu, and W. Yin. A block coordinate descent method for
  regularized multiconvex optimization with applications to nonnegative
  tensor factorization and completion. SIAM Journal on imaging sciences, 6,
  pp.1758-1789, 2013.
}
\seealso{
\code{\link{set_hyper}}, \code{\link{set_init}},
  \code{\link{set_blocks}}, \code{\link{set_groups}}
  and \code{\link{set_struct}}.
}
