% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_hyper_init.R
\name{set_init}
\alias{set_init}
\title{Gather initial variational parameters provided by the user.}
\usage{
set_init(d, p, r, gam_vb, mu_beta_vb, om, s02, s2, sig2_beta_vb, tau_vb)
}
\arguments{
\item{d}{Number of responses.}

\item{p}{Number of candidate predictors.}

\item{r}{Number of candidate annotations.}

\item{gam_vb}{Matrix of size p x d with initial values for the variational
parameter yielding posterior probabilities of inclusion.}

\item{mu_beta_vb}{Matrix of size p x d with initial values for the
variational parameter yielding regression coefficient estimates for
predictor-response pairs included in the model.}

\item{om}{Vector of length r with initial values for the EM hyperparameter 
yielding the probability parameters for the annotation variables.}

\item{s02}{Vector of length 1 with initial value for the EM hyperparameter 
yielding the variance of the predictor-specific modulations of the 
probability parameter for the predictor-response associations (these inital 
values are the same for all predictors, for simplicity).}

\item{s2}{Vector of length 1 with initial value for the EM hyperparameter 
yielding the variance of the annotation effects on the probability 
parameter for the predictor-response associations.}

\item{sig2_beta_vb}{Vector of length d with initial values for
the variational parameter yielding estimates of effect variances for
predictor-response pairs included in the model. These values are the same
for all the predictors (as a result of the predictor variables being
standardized before the variational algorithm).}

\item{tau_vb}{Vector of length d with initial values for the variational parameter
yielding estimates for the continuous response residual precisions.}
}
\value{
An object of class "\code{init}" preparing user initial values for
  the variational parameters in a form that can be passed to the
  \code{\link{epispot}} function.
}
\description{
This function must be used to provide initial values for the variational
parameters used in \code{\link{epispot}}.
}
\details{
The \code{\link{epispot}} function can also be used with default initial
parameter choices (without using \code{\link{set_init}}) by setting
its argument \code{list_init} to \code{NULL}.
}
\examples{

seed <- 123; set.seed(seed)

###################
## Simulate data ##
###################

## Examples using small problem sizes:
##
n <- 50; p <- 60; p_act <- 10; d <- 25; d_act <- 15; r <- 10

## Candidate predictors (subject to selection)
##
# Here example with common genetic variants under Hardy-Weinberg equilibrium
#
X_act <- matrix(rbinom(n * p_act, size = 2, p = 0.25), nrow = n)
X_inact <- matrix(rbinom(n * (p - p_act), size = 2, p = 0.25), nrow = n)

# shuffle indices 
shuff_x_ind <- sample(p)
shuff_y_ind <- sample(d)

X <- cbind(X_act, X_inact)[, shuff_x_ind]

# Association pattern and effect sizes
#
pat <- matrix(FALSE, ncol = d, nrow = p)
bool_x <- shuff_x_ind <= p_act
bool_y <- shuff_y_ind <= d_act

pat_act <- beta_act <- matrix(0, nrow = p_act, ncol = d_act)
pat_act[sample(p_act * d_act, floor(p_act * d_act / 5))] <- 1
beta_act[as.logical(pat_act)] <-  rnorm(sum(pat_act))

pat[bool_x, bool_y] <- pat_act

# Gaussian responses
#
Y_act <- matrix(rnorm(n * d_act, mean = X_act \%*\% beta_act), nrow = n)
Y_inact <- matrix(rnorm(n * (d - d_act)), nrow = n)

Y <- cbind(Y_act, Y_inact)[, shuff_y_ind]

# Annotation variables
#
V <- matrix(rnorm(p * r), nrow = p)
V[bool_x, ] <- rnorm(p_act * r, mean = 2)

################################
## Specify initial parameters ##
################################

tau_vb <- rep(1, d)

gam_vb <- matrix(rbeta(p * d, shape1 = 1, shape2 = 4 * d - 1), nrow = p)

mu_beta_vb <- matrix(rnorm(p * d), nrow = p)
sig2_beta_vb <- 1 / rgamma(d, shape = 2, rate = 1)

om <- rep(0.5, r)
s02 <- 1 / d
s2 <- 0.01

list_init <- set_init(d, p, r, gam_vb, mu_beta_vb, om, s02, s2, sig2_beta_vb, 
                      tau_vb)

########################
## Infer associations ##
########################

res_epispot <- epispot(Y = Y, X = X, V = V, p0 = p0, list_init = list_init, 
                       user_seed = seed)
                       
}
\seealso{
\code{\link{set_hyper}}, \code{\link{epispot}}
}
