% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_epispot.R
\name{set_modules}
\alias{set_modules}
\title{Wrapper of set_blocks to specifically define module partitions.}
\usage{
set_modules(module_ids, module_map = NULL, n_cpus = 1)
}
\arguments{
\item{n_cpus}{Number of CPUs to be used. If large, one should ensure that
enough RAM will be available for parallel execution. Set to 1 for serial
execution.}

\item{tot}{Number of candidate predictors (vector of size 1), and followed 
optionally by the number of responses (vector of size 2).}

\item{pos_bl}{Vector gathering the predictor block positions (first index of
each block), or list gathering the predictor block positions and response
block positions.}

\item{verbose}{If \code{TRUE}, messages are displayed when calling
\code{set_blocks}.}
}
\value{
An object of class "\code{blocks}" preparing the settings for parallel
  inference in a form that can be passed to the \code{\link{epispot}}
  function.
}
\description{
Parallel applications of the method on blocks of candidate predictors for
large datasets allows faster and less RAM-greedy executions.
}
\examples{

}
\seealso{
\code{\link{epispot}, \link{set_blocks}}
}
