% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epispot.R
\name{epispot}
\alias{epispot}
\title{Fit EPISPOT: fully joint & annotation-driven multiple-response regression 
by anneal_scheduleed variational EM inference.}
\usage{
epispot(
  Y,
  X,
  V,
  p0,
  anneal_schedule = c(1, 2, 10),
  list_modules = NULL,
  bin_annot_freq = 0.05,
  list_hyper = NULL,
  list_init = NULL,
  user_seed = NULL,
  tol = 0.1,
  adaptive_tol_em = TRUE,
  maxit = 1000,
  anneal_vbem = TRUE,
  save_hyper = FALSE,
  save_init = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{Y}{Response data matrix of dimension n x q, where n is the number of
samples and q is the number of response variables; Y is centred prior to 
the run.}

\item{X}{Input matrix of dimension n x p, where p is the number of candidate
predictors. No intercept must be supplied; X is scaled prior to the run 
- beware the interpretation of the regression estimates.}

\item{V}{Annotation matrix of dimension p x r, where r is the number of
variables representing external information on the candidate predictors
which may make their selection more or less likely. V is standardised prior
to the run - beware the interpretation of the annotation effect estimates.
Annotations with either concern or do not concern given predictors should 
be coded as binary.}

\item{p0}{Vector of size 2 whose entries are the prior expectation and 
variance of the number of predictors associated with each response.
Must be \code{NULL} if \code{list_init} and \code{list_hyper} are both 
non-\code{NULL}.}

\item{anneal_schedule}{Parameters for annealing scheme. Must be a vector 
whose first entry is the type of schedule: 1 = geometric spacing (default), 
2 = harmonic spacing or 3 = linear spacing, the second entry is the initial 
temperature (default is 2), and the third entry is the temperature grid 
size (default is 10). If \code{NULL}, no annealing is performed.}

\item{list_modules}{An object of class "\code{modules}" containing settings 
for parallel inference using modules of responses. Must be filled using
the \code{\link{set_modules}} function or must be \code{NULL} for no
partitioning into modules.}

\item{bin_annot_freq}{Minimal frequency for binary annotations (if any), 
i.e., annotation variables which concern less than 
\code{bin_annot_freq} x 100 % or more than 
(1 - \code{bin_annot_freq}) x 100 candidate predictors 
are removed prior to the analysis, as they may not be 
sufficiently informative. Default is 0.05. If set to 
NULL, no filter is applied.}

\item{list_hyper}{An object of class "\code{hyper}" containing the model
hyperparameters. Must be filled using the \code{\link{set_hyper}}
function or must be \code{NULL} for default hyperparameters.}

\item{list_init}{An object of class "\code{init}" containing the initial
variational and EM parameters. Must be filled using the 
\code{\link{set_init}} function or be \code{NULL} for a default 
initialisation.}

\item{user_seed}{Seed set for reproducible default choices of hyperparameters
(if \code{list_hyper} is \code{NULL}) and initial variational parameters
(if \code{list_init} is \code{NULL}). Default is \code{NULL}, no
seed set.}

\item{tol}{Tolerance for the stopping criterion.}

\item{adaptive_tol_em}{Boolean indicating whether the tolerance for the 
within-EM variational runs should be controlled adaptively depending on the
EM-convergence status in order to save computational time. Default is 
\code{TRUE}.}

\item{maxit}{Maximum number of iterations allowed.}

\item{anneal_vbem}{Parameters for annealing scheme for the internal runs of 
the variational EM algorithm. Default is geometric spacing, initial 
temperature is 2 and grid size is 10. See \code{anneal_schedule}.}

\item{save_hyper}{If \code{TRUE}, the hyperparameters used for the model are
saved as output.}

\item{save_init}{If \code{TRUE}, the initial variational parameters used for
the inference are saved as output.}

\item{verbose}{If \code{TRUE}, messages are displayed during execution.}
}
\value{
An object of class "\code{epispot}" containing the following output:
 \item{beta_vb}{Estimated effect size matrix of dimension p x q. Entry (s, t) 
                corresponds to the variational posterior mean 
                (mu_beta_vb_st x gam_vb_st) of the regression effect between 
                candidate predictor s and response t.}
 \item{gam_vb}{Posterior inclusion probability matrix of dimension p x q.
               Entry (s, t) corresponds to the posterior probability of
               association between candidate predictor s and response t.}
 \item{xi_vb}{Vector of size r, where entry l contains the estimated marginal
              effect of annotation l on the probabilities of association.}
 \item{rho_vb}{Posterior inclusion probability vector of size r for the
               annotation variables.}
 \item{theta_vb}{Vector of length p containing the posterior mean of theta. 
                 Entry s corresponds to the propensity of candidate predictor 
                 s to be included in the model.}
 \item{zeta_vb}{Vector of length q containing the posterior mean of zeta.
                Entry t controls the proportion of predictors associated
                with response t.}
 \item{converged}{A boolean indicating whether the algorithm has converged
                  before reaching \code{maxit} iterations.}
 \item{it}{Final number of iterations.}
 \item{lb_opt}{Optimized variational lower bound for the marginal
               log-likelihood (ELBO).}
 \item{diff_lb}{Difference in ELBO between the last and penultimate
                iterations. This may be a useful diagnostic information when
                convergence has not been reached before \code{maxit}.}
 \item{p0}{Vector of length two defining the applied sparsity control.}
 \item{rmvd_cst_x, rmvd_cst_v}{Vectors containing the indices of constant
                               variables in \code{X}, resp. \code{V}, removed
                               prior to the analysis. \code{NULL} if no 
                               constant variable removed.}
 \item{rmvd_coll_x, rmvd_cst_v}{Vectors containing the indices of variables
                                in \code{X}, resp. \code{V}, removed prior
                                to the analysis because collinear to other
                                variables. The entry name indicates the
                                corresponding variable kept in the analysis
                                (i.e., that causing the collinearity for the
                                entry in question). \code{NULL} if no 
                                collinear variable removed.}
 \item{rmvd_bin_annot_freq_v}{If \code{bin_annot_freq} is non-NULL, vector
                              containing the indices of variables in \code{V}
                              removed prior to the analysis because they 
                              their frequency is lower than 
                              \code{bin_annot_freq} or higher than 
                              1-\code{bin_annot_freq}.}
 \item{list_hyper, list_init}{If \code{save_hyper}, resp. \code{save_init},
                              \code{TRUE}, hyperparameters, resp. initial
                              variational parameters, used for inference are
                              saved as output.}
}
\description{
Fit EPISPOT: fully joint & annotation-driven multiple-response regression 
by anneal_scheduleed variational EM inference.
}
\examples{
seed <- 123; set.seed(seed)

###################
## Simulate data ##
###################

## Example using small problem sizes:
##
n <- 50; p <- 60; p_act <- 10; q <- 25; q_act <- 15; r <- 10

## Candidate predictors (subject to selection)
##
# Here example with common genetic variants under Hardy-Weinberg equilibrium
#
X_act <- matrix(rbinom(n * p_act, size = 2, p = 0.25), nrow = n)
X_inact <- matrix(rbinom(n * (p - p_act), size = 2, p = 0.25), nrow = n)

# shuffle indices 
shuff_x_ind <- sample(p)
shuff_y_ind <- sample(q)

X <- cbind(X_act, X_inact)[, shuff_x_ind]

# Association pattern and effect sizes
#
pat <- matrix(FALSE, ncol = q, nrow = p)
bool_x <- shuff_x_ind <= p_act
bool_y <- shuff_y_ind <= q_act

pat_act <- beta_act <- matrix(0, nrow = p_act, ncol = q_act)
pat_act[sample(p_act * q_act, floor(p_act * q_act / 5))] <- 1
beta_act[as.logical(pat_act)] <-  rnorm(sum(pat_act))

pat[bool_x, bool_y] <- pat_act

# Gaussian responses
#
Y_act <- matrix(rnorm(n * q_act, mean = X_act \%*\% beta_act), nrow = n)
Y_inact <- matrix(rnorm(n * (q - q_act)), nrow = n)

Y <- cbind(Y_act, Y_inact)[, shuff_y_ind]

# Annotation variables
#
V <- matrix(rnorm(p * r), nrow = p)
V[bool_x, ] <- rnorm(p_act * r, mean = 2)


########################
## Infer associations ##
########################

# Expectation and variance for the prior number of predictors associated with
# each response
#
p0 <- c(mean(colSums(pat)), 10)

# Inference without module
#
res_epispot <- epispot(Y = Y, X = X, V = V, p0 = p0, user_seed = seed)

# Inference with modules
#
module_ids <- sample(c(rep(1, floor(q/2)), rep(2, q - floor(q/2)))) # 2 modules
list_modules <- set_modules(module_ids, n_cpus = 1)

res_epispot_modules <- epispot(Y = Y, X = X, V = V, p0 = p0, 
                               list_modules = list_modules, user_seed = seed)

}
\references{
H. Ruffieux, B. P. Fairfax, I. Nassiri, E. Vigorito, C. Walllace, S. 
Richardson, L. Bottolo. EPISPOT: an epigenome-driven approach for detecting 
and interpreting hotspots in molecular QTL studies, arXiv, 2020.
}
\seealso{
\code{\link{set_hyper}}, \code{\link{set_init}}, 
\code{\link{set_modules}}.
}
